src/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
sqlalchemy-utils
certifi==2024.7.4
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.8
load-dotenv==0.1.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg-binary==3.2.1
pycodestyle==2.12.1
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pyflakes==3.2.0
python-dotenv==1.0.1
python-telegram-bot==21.4
python-telegram-bot[webhooks]
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
uvicorn[standard]
virtualenv==20.26.3
flask[async]
bcrypt
uvicorn
